refactor (Apache Lucene support): move file to utils, add basic spec
